+++ CHANGELOG.md
@@ -10,4 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 * `Union` and `SymmetricDifference`
 * `Memoize`
 
+### Changed
+* `Difference` renamed `ReverseDifference`
+
 ## [0.1.0](https://github.com/coady/iterset/releases/tag/v0.1.0) - 2024-11-15